moves test files to same component directory
moves user forms to components
